requirements.txt
asgiref==3.5.2
cffi==1.15.1
cryptography==38.0.1
cx-Oracle==8.3.0
Django==3.2
django-cors-headers==3.13.0
django-extensions==3.2.1
djangorestframework==3.13.1
MarkupSafe==2.1.1
mysql-connector-python==8.0.30
mysqlclient>=2.0
numpy==1.23.0
pandas==1.4.3
protobuf==3.20.1
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
Werkzeug==2.2.2
